+++ mobile/lib/db/ml/db.dart
@@ -9,8 +9,8 @@ import 'package:path_provider/path_provider.dart';
 import "package:photos/core/event_bus.dart";
 import "package:photos/db/common/base.dart";
 import "package:photos/db/ml/base.dart";
-import 'package:photos/db/ml/db_fields.dart';
 import "package:photos/db/ml/db_model_mappers.dart";
+import 'package:photos/db/ml/schema.dart';
 import "package:photos/events/embedding_updated_event.dart";
 import "package:photos/extensions/stop_watch.dart";
 import "package:photos/models/ml/clip.dart";
